Reimplement the LocationIterator logic
This new logic in the LocationCollector reimplements the logic of the
LocationIterator, but instead of a iterator it returns a simple list;
this makes the whole logic much easier to understand.

Code from the LocationIterator has been re-used 1:1, most notably the
methods "getResourceSuperType". Only some minor renamings have been made
to reflect the changed field names.